chore: update domains spec mcp.json for integration test
For VS Code input, there is no pickString support for multiple selections (e.g. pickManyString or manyOf argument).
Two seeing approaches are (1) pass multi-arg as single arg string and rely on parsing the string input with custom code; (2) create a script to load multi-arg UI.

It seems more appropriate though we leave to argumetn parsing, as array of string, and rely on adding domains on file (no UI for it)

Example if want to add single domain - pickString:
 {
      "id": "domains",
      "type": "pickString",
      "options": [
        "advanced-security",
        "builds",
        "core",
        "releases",
        "repositories",
        "search",
        "test-plans",
        "wiki",
        "work",
        "work-items"
      ]
    }